refactor(direct-control): extract progression read types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves only traditions view input/action/summary/result types and progress dashboard input/legacy-path/result types into packages/civ7-direct-control/src/play/progression/reads.ts. Keeps public facade type re-exports in index.ts and replaces moved progression type imports with concrete session/runtime-probe owner imports.

Preserves public type contracts and does not change traditions or progress dashboard source strings, command serialization, parser labels, read-only/no-send behavior, runtime proof status, telemetry, AI ingestion, hotseat runtime proof, CLI semantic projection, Effect/oRPC procedure-core work, or Task 2.9.4 matrix-row acceptance. Diplomacy/narrative closeout, tactical, operation, ready, public procedure schema, and telemetry type ownership remain pending separate owner slices.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- progression-reads.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.play.progression-read.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle rg scans

Local package/type relocation proof only; no runtime/live-game, progression read behavior, CLI semantic projection, telemetry, AI-ingestion, hotseat product-path, or procedure-core proof claimed.
