+++ packages/civ7-direct-control/src/index.ts
@@ -250,6 +250,14 @@ import type {
 import {
   getCiv7ProgressDashboard as getCiv7ProgressDashboardFromModule,
   getCiv7TraditionsView as getCiv7TraditionsViewFromModule,
+  type Civ7ProgressDashboardInput,
+  type Civ7ProgressDashboardLegacyPath,
+  type Civ7ProgressDashboardResult,
+  type Civ7TraditionAction,
+  type Civ7TraditionActionKind,
+  type Civ7TraditionSummary,
+  type Civ7TraditionsViewInput,
+  type Civ7TraditionsViewResult,
 } from "./play/progression/reads.js";
 import { buildCultureChoiceCloseoutCommand } from "./play/progression/culture.js";
 import { buildTechnologyChoiceCloseoutCommand } from "./play/progression/technology.js";
@@ -400,6 +408,16 @@ export type {
   Civ7UnitSummaryInput,
   Civ7UnitSummaryResult,
 } from "./play/summaries.js";
+export type {
+  Civ7ProgressDashboardInput,
+  Civ7ProgressDashboardLegacyPath,
+  Civ7ProgressDashboardResult,
+  Civ7TraditionAction,
+  Civ7TraditionActionKind,
+  Civ7TraditionSummary,
+  Civ7TraditionsViewInput,
+  Civ7TraditionsViewResult,
+} from "./play/progression/reads.js";
 export type {
   Civ7FullMapGridIdentityCheck,
   Civ7FullMapGridInput,
@@ -582,133 +600,6 @@ export type Civ7ResourceBuilderDiagnosticsResult = Readonly<{
   omittedCells: number;
   resources: ReadonlyArray<Civ7ResourceBuilderDiagnosticsResource>;
   cells: ReadonlyArray<Civ7ResourceBuilderDiagnosticsCell>;
-}>;
-
-export type Civ7TraditionActionKind = "activate" | "deactivate";
-
-export type Civ7TraditionAction = Readonly<{
-  kind: Civ7TraditionActionKind;
-  action: number | null;
-  operationType: "CHANGE_TRADITION";
-  args: Readonly<{
-    TraditionType: number;
-    Action: number | null;
-  }>;
-  validation: Civ7RuntimeProbe<unknown>;
-  cli: string;
-}>;
-
-export type Civ7TraditionSummary = Readonly<{
-  id: number;
-  type: string | null;
-  name: string | null;
-  description: string | null;
-  ageType: string | null;
-  cultureSlotType: string | null;
-  traitType: string | null;
-  isCrisis: boolean;
-  active: boolean;
-  unlocked: boolean;
-  recentUnlock: boolean;
-  actionHints: ReadonlyArray<Civ7TraditionAction>;
-}>;
-
-export type Civ7TraditionsViewInput = Readonly<{
-  playerId?: number;
-}>;
-
-export type Civ7TraditionsViewResult = Readonly<{
-  host: string;
-  port: number;
-  state: Civ7TunerState;
-  playerId: number;
-  turn: Civ7RuntimeProbe<number>;
-  turnDate: Civ7RuntimeProbe<string>;
-  governmentType: Civ7RuntimeProbe<number>;
-  government: Readonly<{
-    type: string | null;
-    name: string | null;
-  }>;
-  slots: Readonly<{
-    total: Civ7RuntimeProbe<number>;
-    normal: Civ7RuntimeProbe<number>;
-    crisis: Civ7RuntimeProbe<number>;
-    active: number;
-    unlocked: number;
-    available: number;
-    open: number;
-  }>;
-  actions: Readonly<{
-    activate: number | null;
-    deactivate: number | null;
-  }>;
-  active: ReadonlyArray<Civ7TraditionSummary>;
-  available: ReadonlyArray<Civ7TraditionSummary>;
-  recentUnlocks: ReadonlyArray<Civ7TraditionSummary>;
-  traditions: ReadonlyArray<Civ7TraditionSummary>;
-  recommendedCli: ReadonlyArray<string>;
-  hiddenInfoPolicy: "player-culture-runtime";
-  notes: ReadonlyArray<string>;
-}>;
-
-export type Civ7ProgressDashboardInput = Readonly<{
-  playerId?: number;
-}>;
-
-export type Civ7ProgressDashboardLegacyPath = Readonly<{
-  legacyPathType: string | null;
-  legacyPathClassType: string | null;
-  ageType: string | null;
-  name: string | null;
-  description: string | null;
-  enabledByDefault: boolean;
-  enabledForPlayer: boolean | null;
-  score: Civ7RuntimeProbe<number>;
-  finalRequiredPathPoints: number | null;
-  nextMilestone: unknown;
-  milestones: ReadonlyArray<unknown>;
-}>;
-
-export type Civ7ProgressDashboardResult = Readonly<{
-  host: string;
-  port: number;
-  state: Civ7TunerState;
-  localPlayerId: number;
-  playerId: number;
-  turn: Civ7RuntimeProbe<number>;
-  turnDate: Civ7RuntimeProbe<string>;
-  age: Readonly<{
-    hash: unknown;
-    ageType: string | null;
-    name: string | null;
-    chronologyIndex: unknown;
-    isFinalAge: Civ7RuntimeProbe<boolean>;
-    isSingleAge: Civ7RuntimeProbe<boolean>;
-    isExtendedGame: Civ7RuntimeProbe<boolean>;
-    isAgeOver: Civ7RuntimeProbe<boolean>;
-    currentAgeProgressionPoints: Civ7RuntimeProbe<number>;
-    maxAgeProgressionPoints: Civ7RuntimeProbe<number>;
-    primaryAgeProgression: Civ7RuntimeProbe<unknown>;
-  }>;
-  player: Readonly<{
-    team: unknown;
-    historicalLegacyPointCountForTeam: Civ7RuntimeProbe<number>;
-  }>;
-  legacyPaths: ReadonlyArray<Civ7ProgressDashboardLegacyPath>;
-  victories: Readonly<{
-    rows: ReadonlyArray<unknown>;
-  }>;
-  triumphs: Readonly<{
-    count: number;
-    rows: ReadonlyArray<unknown>;
-    source: "runtime-gameinfo";
-  }>;
-  proof: Readonly<{
-    victoryManagerGlobal: Civ7RuntimeProbe<string>;
-    sources: ReadonlyArray<string>;
-  }>;
-  hiddenInfoPolicy: "local-player-runtime-progress";
-  notes: ReadonlyArray<string>;
 }>;
 
 export type Civ7SavedGameConfigurationSummary = Readonly<{
